Optimize depth passes (including shadow map generation).
This is done by splitting each vertex buffer in two, one for positions and
one for normals. This way it is possible to use vertex buffers with only
positions in them when running the depth passes. The result is about 10-20%
performance increase. Although, only when having more complex objects - i.e.
not my standard test scene with a lot of instanced cubes.