Introduce a table of materials in the shader.
This removes the need to set a number of root parameters before each
draw call from the CPU. Instead only a single material id constant
need to be set, which is then used in the shader to index into the
materials table. This also makes it easier and more scalable to add
additional texture maps to materials.